Adding new test scatter-partial-failure.js and deleteing the old test
